Move GameState module
